Git_Projects/Smart_Parking_Finder/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0
pydantic==2.5.0
fastapi-mail==1.5.4
